chore: Refactor to to lay groundwork for postgres support
- This was broken off from PR 302 and the postgres details were removed. The refactor was kept, including tests that use macros to generate tests for multiple backends (just one for now).
- c1 lib.rs monomorphizes the run function based on the database (over the 5 store traits that are required at compile time). This makes it more clear where a new type would be added, and how to change an existing type (and illustrates that we may want to use dynamic dispatch for the store).
- Updates the sqlite pool timeouts.
- Adds `put` the bitswap store trait.
