+++ Makefile
@@ -54,13 +54,14 @@ gen-kubo-rpc-server:
 check-kubo-rpc-server:
 	./ci-scripts/check_generated_server.sh kubo-rpc-server ./ci-scripts/gen_kubo_rpc_server.sh
 
-.PHONY: check-queries
-check-queries:
-	./ci-scripts/check_queries.sh "sqlite"
+.PHONY: check-migrations
+check-migrations:
+	MIGRATE_DB=1 ./ci-scripts/check_migrations.sh
 
-.PHONY: check-queries-ci
-check-queries-ci:
-	CI_RUN=1 ./ci-scripts/check_queries.sh "sqlite"
+# Applies migrations to a sqlite without prompting and deletes the file afterward
+.PHONY: check-migrations-ci
+check-migrations-ci:
+	MIGRATE_DB=1 CI_RUN=1 MIGRATION_CLEANUP=1 ./ci-scripts/check_migrations.sh
 
 .PHONY: release
 release:
