--- store/.sqlx/query-1c3b89021ee5b0ce5f81f23176df6f5ffaf544cf6ab93a300a3fc1ec1713aae9.json
@@ -1,20 +1,0 @@
-{
-  "db_name": "SQLite",
-  "query": "SELECT COALESCE(MAX(delivered), 0) as delivered FROM ceramic_one_event;",
-  "describe": {
-    "columns": [
-      {
-        "name": "delivered",
-        "ordinal": 0,
-        "type_info": "Int"
-      }
-    ],
-    "parameters": {
-      "Right": 0
-    },
-    "nullable": [
-      false
-    ]
-  },
-  "hash": "1c3b89021ee5b0ce5f81f23176df6f5ffaf544cf6ab93a300a3fc1ec1713aae9"
-}
