--- store/.sqlx/query-4226e72411f1e37465c4623050156ce38fba3149061ef73f5e2e8f69c525a259.json
@@ -1,12 +1,0 @@
-{
-  "db_name": "SQLite",
-  "query": "INSERT INTO ceramic_one_event (\n                    order_key, cid,\n                    ahash_0, ahash_1, ahash_2, ahash_3,\n                    ahash_4, ahash_5, ahash_6, ahash_7\n                ) VALUES (\n                    $1, $2,\n                    $3, $4, $5, $6,\n                    $7, $8, $9, $10\n                );",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Right": 10
-    },
-    "nullable": []
-  },
-  "hash": "4226e72411f1e37465c4623050156ce38fba3149061ef73f5e2e8f69c525a259"
-}
