--- store/.sqlx/query-727bfcbb45eb6861f3471f74594e96d920f32358195c3e500c5deef9eb83e4ac.json
@@ -1,20 +1,0 @@
-{
-  "db_name": "SQLite",
-  "query": "\n        SELECT\n            order_key as id\n        FROM\n            ceramic_one_event\n        WHERE\n            order_key > $1 AND order_key < $2\n        ORDER BY\n            order_key DESC\n        LIMIT\n            1\n        ;",
-  "describe": {
-    "columns": [
-      {
-        "name": "id",
-        "ordinal": 0,
-        "type_info": "Blob"
-      }
-    ],
-    "parameters": {
-      "Right": 2
-    },
-    "nullable": [
-      false
-    ]
-  },
-  "hash": "727bfcbb45eb6861f3471f74594e96d920f32358195c3e500c5deef9eb83e4ac"
-}
