--- store/.sqlx/query-759328f5a3ccd51b74de5f8c1dceba6f2cb179852607182bf072cc8c38217b06.json
@@ -1,20 +1,0 @@
-{
-  "db_name": "SQLite",
-  "query": "\n        SELECT\n            count(order_key) as cnt\n        FROM\n            ceramic_one_event\n        WHERE\n            order_key > $1 AND order_key < $2\n        ;",
-  "describe": {
-    "columns": [
-      {
-        "name": "cnt",
-        "ordinal": 0,
-        "type_info": "Int"
-      }
-    ],
-    "parameters": {
-      "Right": 2
-    },
-    "nullable": [
-      false
-    ]
-  },
-  "hash": "759328f5a3ccd51b74de5f8c1dceba6f2cb179852607182bf072cc8c38217b06"
-}
