--- store/.sqlx/query-7ef7b58743ad36cef3fc150708c4bc6b503d87cc513a0ea6dae903a0d33ff1d8.json
@@ -1,20 +1,0 @@
-{
-  "db_name": "SQLite",
-  "query": "SELECT\n                order_key as id\n            FROM\n                ceramic_one_event\n            WHERE\n                order_key > $1 AND order_key < $2\n            ORDER BY\n                order_key ASC\n            LIMIT\n            1",
-  "describe": {
-    "columns": [
-      {
-        "name": "id",
-        "ordinal": 0,
-        "type_info": "Blob"
-      }
-    ],
-    "parameters": {
-      "Right": 2
-    },
-    "nullable": [
-      false
-    ]
-  },
-  "hash": "7ef7b58743ad36cef3fc150708c4bc6b503d87cc513a0ea6dae903a0d33ff1d8"
-}
