--- store/.sqlx/query-86c1644963d8037df1a1145daa3017ec70a4a9d2fbe5e30ad70cfc70eb520454.json
@@ -1,26 +1,0 @@
-{
-  "db_name": "SQLite",
-  "query": "SELECT first.id as first, last.id as last\n                FROM\n                    (\n                        SELECT order_key as id\n                        FROM ceramic_one_event\n                        WHERE\n                            order_key > $1 AND order_key < $2\n                        ORDER BY order_key ASC\n                        LIMIT 1\n                    ) as first\n                JOIN\n                    (\n                        SELECT order_key as id\n                        FROM ceramic_one_event\n                        WHERE\n                            order_key > $1 AND order_key < $2\n                        ORDER BY order_key DESC\n                        LIMIT 1\n                    ) as last;",
-  "describe": {
-    "columns": [
-      {
-        "name": "first",
-        "ordinal": 0,
-        "type_info": "Blob"
-      },
-      {
-        "name": "last",
-        "ordinal": 1,
-        "type_info": "Blob"
-      }
-    ],
-    "parameters": {
-      "Right": 2
-    },
-    "nullable": [
-      false,
-      false
-    ]
-  },
-  "hash": "86c1644963d8037df1a1145daa3017ec70a4a9d2fbe5e30ad70cfc70eb520454"
-}
