--- store/.sqlx/query-d4af38b72a8484920733c7177672bacd064fc979b38d448e7c7e686f8c82b166.json
@@ -1,12 +1,0 @@
-{
-  "db_name": "SQLite",
-  "query": "INSERT INTO ceramic_one_event_block (event_cid, idx, root, block_multihash, codec) VALUES ($1, $2, $3, $4, $5) on conflict do nothing;",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Right": 5
-    },
-    "nullable": []
-  },
-  "hash": "d4af38b72a8484920733c7177672bacd064fc979b38d448e7c7e686f8c82b166"
-}
