--- store/.sqlx/query-db3dd73ac37ca2edf7701aa24ed21feb8c0786e47fd764ae4b09556a18077a07.json
@@ -1,12 +1,0 @@
-{
-  "db_name": "SQLite",
-  "query": "INSERT OR IGNORE INTO ceramic_one_root (tx_hash, root, block_hash, timestamp) VALUES (?, ?, ?, ?)",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Right": 4
-    },
-    "nullable": []
-  },
-  "hash": "db3dd73ac37ca2edf7701aa24ed21feb8c0786e47fd764ae4b09556a18077a07"
-}
